Create background.js
This gets the URL when a new page is loaded, stripping the URL of unnecessary stuff and checking if the domain itself actually matches linkedin.com.

If it does, we inject the contents of linkedin.js into the website, which will work to remove our ads. If it doesn’t, we just don’t do anything.